TransactionInput, TransactionOutput: make direct subclass of Message
Since only these 2 classes need to support the concept of a "Parent", it
will be simpler to just add the code to handle that to both of them and
remove the `ChildMessage` altogether.
